refactor: Move key-value-pair out of primitives
