+++ lib/key-value-pair.js
