+++ lib/primitives/base-element.js
@@ -3,7 +3,7 @@
 var _ = require('lodash');
 var uptown = require('uptown');
 var createClass = uptown.createClass;
-var KeyValuePair = require('./key-value-pair');
+var KeyValuePair = require('../key-value-pair');
 
 module.exports = function(registry) {
   return createClass({
