+++ lib/primitives/member-element.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var KeyValuePair = require('./key-value-pair');
+var KeyValuePair = require('../key-value-pair');
 
 
 module.exports = function(BaseElement, registry) {
