+++ lib/serialisers/json.js
@@ -3,7 +3,7 @@
 var createClass = require('uptown').createClass;
 var _ = require('lodash');
 var Namespace = require('../namespace');
-var KeyValuePair = require('../primitives/key-value-pair');
+var KeyValuePair = require('../key-value-pair');
 
 module.exports = createClass({
   constructor: function(namespace) {
